rename interner2.rs -> interner.rs
